requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioredis==2.0.1
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
curl_cffi==0.14.0
dacite==1.9.2
Deprecated==1.3.1
distro==1.9.0
filelock==3.25.2
frozenlist==1.8.0
google-auth==2.49.1
google-genai==1.67.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jaconv==0.5.0
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
nest-asyncio==1.6.0
openai==2.28.0
pillow==12.1.1
propcache==0.4.1
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2026.1.post1
redis==7.2.1
requests==2.32.5
rich==13.9.4
ruamel.yaml==0.19.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
websockets==14.2
wrapt==2.1.2
yarl==1.22.0
yt-dlp==2026.3.3
yt-dlp-ejs==0.5.0
ytmusicapi==1.11.5
fastapi>=0.110
uvicorn[standard]>=0.29
PyJWT>=2.8
pydantic>=2.0
